Restructure
Split the code into elvind and elvin (lib)

Fork out xdr code from packets

Add the start of some xdr testing

Introduce a connrply packet without a suitable abstraction for now.

For local compilation this requires setting of GOPATH to rootdir (to get library).

In turn this requires the introduction of a top-level src dir as GOPATH expects this in the layout