Add range checking and error handling to the XdrPut* family.
This also means we're stricter about int vs int32 so push that change all
the way through.

Fix a readyBytes bug where we could be out of bounds.

Separate the closer out from the reader/writer (future plan).

Add a couple of wee (but informative) benchmarks for xdr which can be run
with: 'go test -bench Bench'