Introduce Nack which reorganizes the packet queues.
Subscribe can return SubRply, Nack or Disconn so we need to be able to
return one of three different packet types. Rework the sbscribe channel
so it deals with the Packet interface type across underling packet types.

Connect can also return Nack so track XIDs on a packet channel for that
too.

Disconnect doesn't Nack but give it a packet channel for consistency.

Now we're dealing with Nack we need to deal with Elvin's positional
arguments and format strings so add some support for that.