Add Disconnect (Rqst/Rply and Disconn).
This requires some work on the state machine to cater for cleaning up so we can begin
to ponder reconnect and redirect.

Introduce a top level client.Close() to make a clean state from which we will resubscribe
which cleans up the previous simplistic socket closure.

Introduce a wait group so we can synchronise the read and write handler goroutine exits.

Add some (interim) server hacks for testing.